requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
Werkzeug
pandas
tensorflow 
pillow 
numpy
gunicorn==21.2.0
